Code Review: Move DataResultModels locations
Code Review: switch statement
