+++ src/main/java/bio/terra/service/search/SearchService.java
@@ -5,9 +5,9 @@ import bio.terra.model.SearchIndexModel;
 import bio.terra.model.SearchIndexRequest;
 import bio.terra.model.SearchQueryRequest;
 import bio.terra.model.SearchQueryResultModel;
-import bio.terra.service.filedata.google.bq.BigQueryDataResultModel;
 import bio.terra.service.search.exception.SearchException;
 import bio.terra.service.snapshot.Snapshot;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryDataResultModel;
 import bio.terra.service.tabulardata.google.bigquery.BigQuerySnapshotPdao;
 import java.io.IOException;
 import java.util.ArrayList;
