+++ src/test/java/bio/terra/service/tabulardata/google/bigquery/BigQueryPdaoUnitTest.java
@@ -40,7 +40,6 @@ import bio.terra.service.dataset.DatasetTable;
 import bio.terra.service.dataset.GoogleStorageResource;
 import bio.terra.service.filedata.exception.TooManyDmlStatementsOutstandingException;
 import bio.terra.service.filedata.google.bq.BigQueryConfiguration;
-import bio.terra.service.filedata.google.bq.BigQueryDataResultModel;
 import bio.terra.service.resourcemanagement.google.GoogleProjectResource;
 import bio.terra.service.snapshot.RowIdMatch;
 import bio.terra.service.snapshot.Snapshot;
